Share Vite plugins across package builds
Centralize compressed JSON loading and Three.js compatibility aliases so root and extension builds use one tested implementation.
